+++ docs/changes.rst
@@ -1,6 +1,13 @@
 Changes
 =======
 
+Version 0.2.3, 2020-05-02
+
+* Fixed :py:exc:`ForeignKeyViolation` when building normalized temporary table
+  ``characters_to_character``.
+* Fixed :py:exc:`ValueError` when no command was specified for the
+  :command:`pimdb` command line client.
+
 Version 0.2.2, 2020-04-26
 
 * Fixed :py:exc:`AssertionError` when command line option ``--bulk`` was less